front: rtk migration of submitConfAddTrainSchedules.ts
